refactor(schema): Remove enums directory and Update import paths for enums
