+++ packages/schema/src/api-key.ts
@@ -1,5 +1,5 @@
 import z from 'zod'
-import { expiresAfterEnum } from '@/enums/enums'
+import { expiresAfterEnum } from '@/enums'
 
 export const CreateApiKeySchema = z.object({
   name: z.string(),
