+++ packages/schema/src/workspace-role.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod'
-import { authorityEnum } from '@/enums/enums'
+import { authorityEnum } from '@/enums'
 
 export const CreateWorkspaceRoleSchema = z.object({
   name: z.string(),
